translation/requirements.txt
neo4j
python-dotenv
streamlit